# Improve performance Kernel#Float with using Primitive.mandatory_only? method