@ko1 2021-11-17
I'm not sure why they are improved.

```
                       float_true
          built-ruby:  21317416.4 i/s
        compare-ruby:  10875475.8 i/s - 1.96x  slower

                      float_false
          built-ruby:  22846698.6 i/s
        compare-ruby:  13910140.5 i/s - 1.64x  slower
```