@k0kubun 2021-11-18
Here's a result on my environment:

```
$ benchmark-driver benchmark/kernel_float.yml --rbenv 'before;after'
Calculating -------------------------------------
                         before       after
               float    38.387M     42.598M i/s -     10.000k times in 0.000261s 0.000235s
          float_true    26.208M     26.495M i/s -     10.000k times in 0.000382s 0.000377s
         float_false    25.153M     24.703M i/s -     10.000k times in 0.000398s 0.000405s

Comparison:
                            float
               after:  42597783.2 i/s
              before:  38387421.2 i/s - 1.11x  slower

                       float_true
               after:  26494838.8 i/s
              before:  26208197.9 i/s - 1.01x  slower

                      float_false
              before:  25153499.2 i/s
               after:  24702703.0 i/s - 1.02x  slower
```
This looks like what we're looking for.